Cleaned remaining .o files + Changed concatStrings signature

+++ Userland/UserModule/standard_library/include/mystdlib.h
@@ -7,4 +7,4 @@ long strToNum(char* string);
 size_t strToNumPos(char* string);
 int strcmp(char* s1, char* s2);
 size_t strlen(char *string);
-int concatStrings(char strings[][MAX_PARAMETER_LENGTH], int stringAmount, char* output);
+size_t concatStrings(char strings[][MAX_PARAMETER_LENGTH], size_t stringAmount, char* output);
