+++ Userland/UserModule/standard_library/mystdlib.c
@@ -19,8 +19,8 @@ size_t strlen(char *string){
 }
 
 // Concatenates all the strings sent and returns output string length on finalOutputLen
-int concatStrings(char strings[][MAX_PARAMETER_LENGTH], int stringAmount, char* output){
-    int outputIdx=0, strIdx=0, i=0;
+size_t concatStrings(char strings[][MAX_PARAMETER_LENGTH], size_t stringAmount, char* output){
+    size_t outputIdx=0, strIdx=0, i=0;
     while(strIdx < stringAmount){
         if(strings[strIdx][i] == 0){                // *(*(strings+strIdx*stringMaxLen)+i)
             if(strIdx < stringAmount - 1)
